fix build failure with nagfor7.0 because of using "complex" function
"complex" function which is not a fortran standard causes error like
below in nagfor.

"Error: test_linalg.f90, line 445: Implicit type for COMPLEX in TEST_TRACE_CQP"

So, this is replaced with "cmplx" fortran standard function.
At the same time, I specified the "kind" value so that it has the
same precision as the "cmplx" function argument.