fix the arguments of inquire routine
The Fortran standard says "The value of a file-unit-number shall
identify a valid unit.".
Once the newunit value generated by open(newunit=..) is not
connected to a unit, it is not valid.
When the unit number is not valid, the compiler can do anything
from return .TRUE., .FALSE., raise an error. Therefore, I added the iostat
argument for the inquire routine.