Changed test_sorting.f90 to reduce run times
Changed sorted integer array from size 2**20 to size 2**16
Changed sorted character array from size 26**4 to 16**4, so the four characters
are a-p.
Changed sorted string array from size 26**3 to 16**3, so the four characters
are a-p.
Change format for the run times from F10.5 to F10.6 to give at least three
digits of precision.

[ticket: X]