Added parameters test_power and char_set_size
The code had two contexts in which it used the literal value 16.
First for power of two number of elements in the integer arrays.
Second for the number of characters in the character set used to generate
character arrays and strings. For the first conext I defined the parameter
test_power. For the second context I defined the parameter char_set_size,
bot with the value of 16.

[ticket: X]