Added a compiler option
The hash functions require integer values expressed as hexadecimals that map
to negatve intgers under normal unsigned integer to twos complement signed
integer conversions. This is undefined under the Fortran Standard but, is the
default for gfortran 10 and 11 and all recent versions of ifort, but requires
the compiler flag, "-fno-range-check" for gfortran 9 and earlier.

[ticket: X]