Modified Makefile.validation and README.mb.
I attempted to make the validation also possible for the Intel compilers, but
then discovered that one of the C codes assumed that the compiler was either
gcc or MSVC, and could not be compiled by icc. I backed out most of the changes
in Makefile.validation, and documented the limitation in README.md.

[ticket: X]